Inhibit inlining of overflow/underflow per-CPU paths.
With inline restartable sequences (initially introduced in
bd50649c29727d865217de63c20d78e1cbc30610), the compiler can now see the
function we are calling.  While the cold path is not taken *too* often, it may
be hot enough that the compiler inlines the callees, requiring ::operator
delete to construct a stackframe and bloating the function body.

PiperOrigin-RevId: 296489345
Change-Id: Ide27f0eb41420250d836466ab400f052c1e329bb