Introduce partially released huge page list.
The partially released huge page list will contain huge pages which contain a
mix of allocated, free (but not returned), and returned pages.

Currently, bookkeeping is simplified by fully releasing memory when it is
returned to an already released huge page.  This allows us to track the free
bitmap plus a single boolean for whether the entire huge page has been released
or not.

The change to populate this list will be in a subsequent commit and guarded by
the value of partial_rerelease_.

PiperOrigin-RevId: 300643907
Change-Id: I56820a3677b55579e426d0a9755769c89f1af290