Calculate the number of previously unbacked pages when allocating.
The HugePageFiller previously used the huge page's state (was_released) to
determine whether the entirety of the request was previously unmapped for
accounting purposes.  With the commit sequence to retain released pages, this
all-or-nothing state will no longer hold.  Push the determination into the
PageTracker where we can (in a subsequent commit) use a fine-grained bitmap.

PiperOrigin-RevId: 301179450
Change-Id: I2474e72957c8a80e5ca171cea129911b826e1880