Add test to measure CPUCache usage with lazy CPUs.
This introduces a testing path on Activate to avoid setting up the full
fastpath to start using per-CPU caches.  This keeps memory allocations as part
of the test/background threads from perturbing things.

Since we may fault an entire huge page, the bounds are wide.  In the future, we
can improve this by MADV_NOHUGEPAGE'ing the memory to get a more precise
accounting.

PiperOrigin-RevId: 301431426
Change-Id: I51ecccdf0e7f8393583f015eb4946a78790d7a46