Prefetch the most commonly used fields in Span, instead of base pointer.
The size of Span is 48 bytes, which means that 50% of the time the Span is in
two cachelines. 25% of the time, the first 16-bytes are in a different cache
line and 25% of the time the last 16-bytes are in a different cache line.

Previously, the base pointer of Span* was fetched, but the most commonly used
fields are in the middle cacheline. So 25% of the time, the wrong cache line
was prefetched.

PiperOrigin-RevId: 309725165
Change-Id: I0670009a7e706a53a26b05091a4dd907d81a15da