Leverage asm goto for delete fast path.
This was 67d05358ff3a, reverted in a275def91773 due to an LLVM crash.  This is addressed by https://github.com/llvm/llvm-project/commit/41ba80182cc2540fcf85bcda10bbde04816b7e6e.

This removes the need to compute the overflow boolean and compute a second branch on its value.  In microbenchmarks, this branch is typically well-predicted, but this change helps reduce the dynamic instruction count on the fastpath.

PiperOrigin-RevId: 309948818
Change-Id: I3a04c2b958387f611e0a2827ad9c64dde15afa27