Deallocate sampled objects into the respective frontend cache.
In production profiles, ~1/3 of the time in handling sampled objects is spent
deallocating into the TransferCache.

*  Prior to a43cd6635d4e, this intentionally triggered a miss in the
   TransferCache so we'd always write-back to the CentralFreeList.
*  When objects_per_span > 1, we set obj=nullptr.  Therefore we only invoke
   this on single object Spans.  This means the CentralFreeList pushes the
   entire Span into the page heap.  This contributes to pageheap_lock
   contention.

This has the subtle effect of causing a net shift single object span-sized
objects out of the per-CPU/per-thread caches when sampling occurs.

When we are not tracking hit/miss rates, we can return the object to the
per-CPU/per-thread cache.  This is much faster due to implementation, avoiding
many tiers of caches, and locking.

PiperOrigin-RevId: 316105602
Change-Id: Ib0a0e135ff40bed1f8169e4c4b85812a121cf36c