Fix wraparound behavior in LockFreeTransferCache
This changes the LockFreeTransferCache to initialize its counters close to the
point where they would wrap around uint32_t max. This triggered a bug in that
the size calculation was incorrectly handling wraparound. The sequence counters
are in an "unmasked" space, so special handling of wraparound shouldn't be
necessary.

With the change to init the sequence numbers close to uint32_t max, the tests
fail. With the fix to size_from_pos(), the tests pass again.

PiperOrigin-RevId: 340721942
Change-Id: Idb20a1ffe27a746d776b5d45ba89e65e16800793