Optimize for HugePage not overflowing.
The shift left (by kHugePageShift) and shift right (by kPageShift) needs to
ensure the upper bits are cleared.  The upper bits would only be set if we
created an invalid HugePage that did not correspond to an addressable pointer.

Add assertions that HugePage does not overflow and condense the
kHugePageShift-kPageShift arithmetic into HugePage::first_page().

PiperOrigin-RevId: 344105792
Change-Id: Ia6cf5137a2c175ff00afb7c3aec90b6cf35e3b7d