Centralize the choice of type for compact size class representation
Size classes are mostly represented as uint32_t values throughout TCMalloc, but
there are some cases in which we want to use the most compact representation
possible. Currently those use a single byte, either a uint8_t or an unsigned
char.

In order to introduce NUMA awareness we plan to duplicate the set of size
classes for each (set of) NUMA node(s) - i.e. add at least one to the
multiplier used to derive kNumClasses from kNumBaseClasses. This will cause
some of our configurations to have enough size classes that they can no longer
be represented by a uint8_t. For example, with 8KiB pages (i.e.
TCMALLOC_PAGE_SHIFT == 13) we have kNumBaseClasses == 86. If we double this to
represent just two (sets of) NUMA node(s) and have the existing set of extended
size classes then we have a kNumClasses multiplier of 3, giving us kNumClasses
== 258 which exceeds the range of values that a uint8_t can represent. For
256KiB pages kNumClasses is higher still, and the value will scale with the
number of distinct sets of NUMA nodes that we want to support.

Here we prepare for kNumClasses exceeding 256 by centralizing the choice of the
type used for compact representation of a size class, introducing a new
SizeClass type alias for the appropriate integer type. Code which explicitly
used uint8_t or unsigned char to represent a size class is adjusted to use the
new SizeClass alias & will naturally adjust to use a uint16_t when kSizeClasses
is too large for a uint8_t. We could in theory use SizeClass more widely to
replace current uses of uint32_t, but that is left as a potential exercise for
the future.

PiperOrigin-RevId: 378896003
Change-Id: I3da82ddd458b096c210116498ea08cb6210ad274