Add tcmalloc_numa_aware build target
Introduce a tcmalloc_numa_aware build target which builds TCMalloc with NUMA
awareness included. This enables support for 2 NUMA partitions, which is all we
currently support. This could be extended in the future if needed.

Note that this only enables the support at compile time - i.e. we compile in
NUMA awareness when using this build target, but it is still disabled at
runtime by default. NUMA awareness can be enabled at runtime for a program
which uses tcmalloc_numa_aware as its malloc either by introducing a dependency
for that program upon the want_numa_aware BUILD target, or by setting the
TCMALLOC_NUMA_AWARE environment variable before starting the program.

PiperOrigin-RevId: 380017932
Change-Id: I412117a0adec8a9e86bcbfeeeb4266681914ea39