Store scaled NUMA partition numbers
We currently store raw NUMA partition numbers in the NumaTopology
cpu_to_partitions_ array, and then in the fast path when we look up the
partition number we have to multiply it by kNumBaseClasses to obtain the value
we should add to our size class.

Adjust NumaTopology to store partition numbers pre-multiplied by
kNumBaseClasses instead, which allows us to avoid the multiplication in the
fast path.

PiperOrigin-RevId: 381286543
Change-Id: If6a15cad7cc584abf8fb2d3ca8caa1e08eea2dad