Don't stop refilling when receiving a batch smaller than batch size.
Then new ring-buffer based transfer cache will return only as many items as are
in the transfer cache in one call to RemoveRange(). Right now, the same happens
when RemoveRange() hits the free list: if the free list contains fewer than N
items, it doesn't repopulate by default. With this change, we would force it to
repopulate.

Update one test to use a new allocation size. Looks like this change made it so
we cache more items of that size in the per-cpu cache, which means we wouldn't
hit the system allocator for this test.

PiperOrigin-RevId: 384786213
Change-Id: I1409980eb48a75009da97f7e7cccb77da6c8695c