Fix missing update of low_water_mark_ when inserting.
This is required e.g. when inserting a single item into a full TF cache. We
return one batch to the freelist but only re-insert one item, thus reducing the
final length of the TF.

Add support for plundering to the fuzz test and add more asserts which helped
with debugging this.

PiperOrigin-RevId: 416291238
Change-Id: I1a500d5892ce3f9e92727ab2a22b4019c09888ae