Assert that rseq TLS data has been registered during Push/Pop.
This invariant is maintained by the TCMalloc fast path, but during testing of
individual components of CPUCache/TcmallocSlab, we may access these structures
from a thread that has not registered itself with the kernel.

This can lead to an out-of-bounds access (cpu_id = -1) which is harder to debug
than an assertion failure.

PiperOrigin-RevId: 417701134
Change-Id: I0f4916ef718bce6f582a1c586b991e46a6dc24f4