Estimate internal fragmentation of sampled allocations.
TCMalloc allocates memory in discrete size classes, so a request for N bytes
may be rounded up to M>N if N is not a size class of its own.  We can estimate
this cost on sampled allocations.

This information is available by iterating over the current heap profile, but
we can maintain a running total for easy reference.  The newly added tests
confirm that the result of GetNumericProperty is consistent with the heap
profile-based method.

PiperOrigin-RevId: 423921514
Change-Id: I9796f7627267e2dd4656bf0c17aab416cc58d117