Don't rely on the return type of absl::bit_width.
Because absl::bit_width sometimes is just an alias for std::bit_width, the
return type may be int or size_t depending on the standard being used with the
compiler and whether the version of the standard library being used implements
the proposal in https://cplusplus.github.io/LWG/issue3656. This causes std::min
to choke due to deducing conflicting types.

PiperOrigin-RevId: 437257599
Change-Id: Iaf2a7ccbcc243cb2576f4067991211178f84fc3f