Fix the bug in which we can Pop from an MADVISE_DONTNEEDed slab in a way that doesn't require changing RSEQ.
Similarly to how we change current and fence the CPU before changing the rest of the header in TcmallocSlab::Drain, we can first change current to 0 and fence all CPUs before MADVISE_DONTNEEDing away the slabs.

This means we don't need to change PopBatch and we can roll back the change to the Push/Pop RSEQ.

Also change `cpu` loop variables to unsigned to silence ClangTidy complaints.

PiperOrigin-RevId: 444280347
Change-Id: Icaf85b94d7d9e5b753f70df933850c2cf78e6615