heap profiling: provide sampled_resident_size
Plumb through sampled_resident_size, which is a mincore call on an
arbitrary pointer for each stack frame corresponding to a sampled
object.

This is a very rough approximation of the resident size, but is better
than nothing. We intend on upgrading this to resident_size when future
improvements in merging stack traces outside of the page heap lock land.

PiperOrigin-RevId: 446462100
Change-Id: I8b82a49e38d14566085dd1f153543a9b0d4bee5b