Fixes a bug in TryPlunder in legacy transfer cache.
TryPlunder tries to remove a batch of objects from the transfer cache when if contains fewer objects. This isn't technically a bug in current setting because we are always guaranteed to have a batched number of objects in the legacy transfer cache (if the cache is non-empty) as it is not flexible. This would change once partial updates are enabled.

PiperOrigin-RevId: 465661666
Change-Id: Iae5a45f39c3af77860688034c4e16f4c40ac0abf