Add malloc_tracing_extension::GetAllocatedAddressRanges that returns the list of address ranges corresponding to Spans allocated by TCMalloc.
To get this list, we walk the PageMap to find the allocated Spans. We do this efficiently using the added get_next_set_page routine which allows us to reduce the iteration space in the 2- or 3-level PageMap.

It is important to not have any allocation event during this walk, as we're holding the pageheap_lock. To ensure this, space is reserved in the output vector prior to the walk.

PiperOrigin-RevId: 469872432
Change-Id: Icc3bac153b4d4a6fba3e2b88daaac6561b26adb7