tcmalloc: fix building with -DTCMALLOC_PERCPU_USE_RSEQ=0
It may be useful to build tcmalloc w/o rseq use for reasons.

TCMALLOC_PERCPU_USE_RSEQ is defined as:

  #if !defined(TCMALLOC_PERCPU_USE_RSEQ)
  #if ...
  #define TCMALLOC_PERCPU_USE_RSEQ 1
  #else
  #define TCMALLOC_PERCPU_USE_RSEQ 0
  #endif
  #endif

Which suggests that it was intended that a user can pass
at least -DTCMALLOC_PERCPU_USE_RSEQ=0.

However, currently the build is broken for 2 reasons:
1. Undefined reference to __rseq_abi in inline asm.
2. Unused slabs and shift variables.

Repair building with -DTCMALLOC_PERCPU_USE_RSEQ=0.

PiperOrigin-RevId: 471236288
Change-Id: I321630b1be59a06a5dccfa21803c652f742e4dda