Fixes bug in TransferCache's TryPlunder.
We update low water mark, both in TryPlunder and ShrinkCache, so that it maintains lowest value of used bytes in the cache. The test makes sure that we are plundering the right amount during each plunder operation. We also protect accesses to low_water_mark_ with a lock, and remove atomic accesses to it.

PiperOrigin-RevId: 479380866
Change-Id: I1efdbf20212113f7c162ab682e185b8723d1d8b4