Modify donated page stats to keep a net count of donated pages in the filler.
We currently use "slack" to decide when to use the HugeRegions code, but this
only considers large allocations that are still alive.  Over time, large
allocations can be donated to the filler and then partially used.  This also
contributes to space in the filler that must be populated with small
allocations to avoid fragmentation.

We currently only decrement the donated running counter if we successfully
reassemble with the parent allocation--which was useful for virtual address
space contiguity reasons--but this modification would show how many huge pages
in the donated state *or not* are in the filler because of this.

PiperOrigin-RevId: 484109193
Change-Id: If57bf27b476120922a68af945ea6403ac80a7cf1