Use huge region more often.
When the experiment TEST_ONLY_TCMALLOC_USE_HUGE_REGIONS_MORE_OFTEN is enabled, we use number of abandoned pages to determine when to use HugeRegion. We also relax the requirement for enabling region. If the binary is reasonably sized (i.e. abandoned pages > 64MB), we use HugeRegion for large allocations.

PiperOrigin-RevId: 488501786
Change-Id: Iaef655bf2c4f67a0cf90281aa139b51345505cf8