tcmalloc used compiler-builtins for prefetches.
The compiler builtin __builtin_prefetch(..., /*rw=*/1) is a prefetch
with intent to write. Compilers generate PREFETCHW on x86 only when -march=
broadwell or newer; unfortunately this is not commonly set, leading to a
prefetch-with-intent-to-read being emitted instead.

Add a wrapper for prefetch functions to encapsulate the compiler builtins.
The PrefetchW wrapper emits PREFETCHW regardless of -march settings; this is
safe on older CPUs as the instruction has been a no-op on Intel CPUs that
don't support it since at least Core 2.

PiperOrigin-RevId: 490023882
Change-Id: I16e1818395b22b33aa29bbfa9cc2f314beed7aa4