tcmalloc/internal/residency_test: change mmap_hint for tsan
residency_test.cc hardcodes an mmap address when ThreadSanitizer
is enabled. The mmap_hint lies within the heap region assumed by
ThreadSanitizer for x86, but it does not lie within the heap (or
app) regions assumed by ThreadSanitizer for ARM.

This CL moves the mmap_hint to the low app region; it has
been tested to work on both x86 and ARM, and theoretically works
(untested) on other platforms such as PowerPC and System 390.

PiperOrigin-RevId: 499377506
Change-Id: I7d9d54c1d296e38ecb081de03e8f222c4fde252a