Adding an experimental feature for calculating demand requirement when making skip subrelease decisions in HugePageFiller.
Introduced a pair of parameters, filler_skip_subrelease_short_interval() and filler_skip_subrelease_long_interval(), for calculating the recent demand requirement. The former finds the recent short-term demand fluctuation peak and the latter finds the recent long-term demand trend. The feature is enabled by configuring either one of the two intervals. The demand requirement is the sum of the two (fluctuation peak and demand trend) if both are configured.

Note, we recommend configuring both intervals, and the short-interval should be (significantly) shorter or equal to the long-interval to avoid realized fragmentation caused by non-recent (short-term) demand spikes.

The feature is not enabled by default: HugePageFiller prefers to use recent demand peak for making skip-subrelease decisions if the existing parameter filler_skip_subrelease_interval() is configured.

PiperOrigin-RevId: 509328699
Change-Id: I734c1f5bd7b8352fadd0fc739cd144e969a4adc3