Use object miss count for background resizing of transfer caches.
This cl introduces insert_object_misses and remove_object_misses that are measured as follows: It counts object misses instead of number of request misses. That is, if the transfer cache misses a batch-sized insert request, the miss count goes up by a batch_size instead of 1. When partial updates are allowed, this helps us differentiate between misses due to say 32-object vs. 16-object requests.

We use object_miss counts to determine which size classes to grow and which ones to shrink.

PiperOrigin-RevId: 511556453
Change-Id: I07c4f479cf737d378773d12f36dcff6bde61db30