Release free pages from partial allocs in the page heap.
We periodically try to release all the free pages in partially-released lists if this feature is enabled and malloc_release_bytes_per_sec is non-zero. The release occurs periodically as a background op. If tcmalloc_skip_subrelease_interval is set as well, we consider realized fragmentation when releasing the pages. We use skip subrelease mechanism to compute the peak demand and make sure that we do not drop below it by releasing the free pages. If skip subrelease mechanism is disabled, we try to release all the free pages in partial allocs to the system. This mechanism may be enabled only for hugepage-aware allocator. It is a no-op for legacy page heap.

This feature is disabled by default and may be enabled by setting --tcmalloc_release_partial_alloc_pages to true.

PiperOrigin-RevId: 518087546
Change-Id: I8f340bf67bbc643cb54231fcc0d6752d9f876cf9