Add an opt-out for generic sharded transfer cache.
We disable sharded transfer cache if runtime control TCMALLOC_GENERIC_SHARDED_TRANSFER_CACHE_CONTROL is set to 0 or the binary is linked to :want_disable_sharded_transfer_cache_test_helper. If either of the two is set, we disable generic sharded transfer cache. Currently, generic sharded transfer cache is disabled by default. We will enable it in a follow-up cl.

PiperOrigin-RevId: 523261517
Change-Id: Ibe6b987be94c1a3a00a1927379b22b1e50389838