Implement improved guarding algorithm in TrySampleGuardedAllocation.
This functionality is guarded by the improved_guarded_sampling feature.

When enabled, the algorithm operates in 2 modes depending on the current rate
of guarding.

(mode: striving) When the rate of guarding is well below (2x) the configured
rate, every sample stack trace will be examined and depending on the number of
guards already present on the stack:
  - >= 4: do not guard
  - == 3: guard again 1% of the time
  - == 2: guard again 10% of the time
  - == 1: guard again 20% of the time
  - == 0: guard again 100% of the time

(mode: passive) Do not guard anything until the rate has fallen sufficiently to
switch back into striving mode.

While the above mentions a lower guarding rate, the code looks at the rate as a
ratio of sampling over guarding, which means that when the guarding rate falls
the ratios mentioned in the code are larger.

PiperOrigin-RevId: 530701699
Change-Id: Ib9a10da260e8252ea00ce479d79a1c6eddf643be