Remove optimization that 'reconstructs' cpu value from slab, shift and scratch registers.
Previously the code took the value of 'slabs' and 'shift' along with the scratch register holding the computed pointer to infer the used cpu. As we are now loading slabs and shift inside the RSEQ, the observed values outside the RSEQ may have changed leading to a bad cpu value. Since this is the slow path executed where a indirect load from vcpu is negligible, and also as this is only relevant on compilers supporting asm goto with register preservation (and thus poorly covered in existing tests), we should simply reload vcpu.

PiperOrigin-RevId: 530750951
Change-Id: I388c92534dc79503f7758581cc46a2104d680dd5