Fix failed test due to enabling separate allocs in the hugepagefiller
The test fails since it can generate an allocation where the size is larger
than that supported by the hugepagefiller and then a portion is donated to the
filler.  If the number of objects is greater than the limit used for deciding
whether or not to use the  many objects alloc, then a particular assert in the
filler fails.

PiperOrigin-RevId: 530971633
Change-Id: I2fbe749e68513e09c8dd79a9d23f12bfb32fd377