tcmalloc: fix realloc performance regression
The regression was introduced in
"tcmalloc: detect UAF and more OOBs on realloced blocks".

Sampled allocations are reallocated and copied even if not strictly
necessary. This is problematic for very large allocations, since some old
programs rely on realloc to be very efficient (e.g. call realloc to the
same size repeatedly assuming it will do nothing). Very large allocations
are both all sampled and expensive to allocate and copy, so don't
reallocate them if not necessary.

The regression is spotted on DTC:
https://git.kernel.org/pub/scm/utils/dtc/dtc.git
It can realloc very large block to the same size
repeatedly while reading in large files.

PiperOrigin-RevId: 541973978
Change-Id: Ia4a3be48a402bf6879f35bfb9018af1a9d49a02a