Pass lifetime information from central freelist to HPAA.
This is a no-op as far as the lifetime-based allocation from HPAA is concerned. We move the logic that determines if the span is long-lived vs. short-lived to CFL, as CFL should have more information about lifetime during span allocation. In the future, we can build smarter heuristics in CFL to pass lifetime information, without needing to pass each signal separately to HPAA.

PiperOrigin-RevId: 547810813
Change-Id: If4d0661ac1f58b3fbe0d7b73fca403437945f0db