tcmalloc: batch MapObjectToSpan
MapObjectToSpan is a tiny function that used to be inlined
when all CFL code was in the source file.
When most of the CFL code was pulled to the header file,
it become outlined (not visible during template instantiation).

It's not trivial to move it to the header file since it requires
definition of tc_globals and causes cyclic dependencies.

Make it operate on a batch of objects to amortize call overheads.

PiperOrigin-RevId: 561849162
Change-Id: I2e3f2cf5cd7a5268a9a688173a2030abdad2f1ab