tcmalloc: fix cpu cache overflow batch size
Currently we use following formula to compute number of object to transfer
between cpu cache and transfer cache for limited capacity caches:

    target = std::max<size_t>(1, (capacity + 1) / 2);

It yields 1 for capacity 2, which is wrong.
1 is used for the overflow/underflow object itself and we don't
refill/drain the cpu cache itself at all. Which means on the next
push we overflow again and don't use the cpu cache capacity to our
benefit at all.
For capacity 6, we will refill/drain only 2 objects, while should 3.

For capacity 1 it also yields 1, which is more controversial but probably
also wrong (series of overflows/underflows are common).

It also rounds to transfer batch size, which leads to addition issues.
For capacity 7, we drain/refill the whole cache, which looks wrong.
Then for capacity 8-9, we drain/refill almost the whole cache.
Then on capacity 24, number of refilled object suddenly jumps from 8 to 16.
With the continuous transfer cache, rounding to batch size is not really
necessary, TC can accept/remove 3 or 7 objects just as well.

Use the following formula instead:

    target = (capacity + 1) / 2 + 1;

It yields 2 for capacity 2 and 4 for capacity 6.
See the test changes for other numbers.

PiperOrigin-RevId: 561876171
Change-Id: I0b88571d80f67a60fa2062a673000975fb317d9a