tcmalloc: tune transfer cache behavior on incomplete batches
Currently we use transfer cache only if we can satisfy the whole transfer.
This looks suboptimal and unnecessary. Say, if we want to transfer 60
objects and have 40, it looks worth using the TC. Especially for insert
operation, since the alternative is sorting objects to spand in CFL,
which is expensive for each object.
Relax the condition to use TC if we can satisfy 2/3 of the transfer.

Also if we decided to use TC for insert during speculative check,
but then when we locked the mutex the capacity has changed,
currently we gave up on using TC. This also looks suboptimal
since already payed all overheads (locked the mutex), so no point
in not using any remaining TC capacity.

PiperOrigin-RevId: 561904260
Change-Id: I093f43e82063c9195da47839e6f72f79084e7bb1