trivial: tcmalloc sampler: replace "p" with "T"
this is a terrible variable choice and is making documentation hard. for one, it
already clashes with our probability density function p(x). for another, p is
the probability that a byte is sampled = 1/T.

in the code itself it's always called sample_period_ so there is no ambiguity
or code change.

PiperOrigin-RevId: 562816706
Change-Id: I456afa9835296374a4335e21974d813e85e48fe0