tcmalloc: don't allocate numa node array if numa is not enabled
Currently we always allocate uint64_t[CPU_SETSIZE] array for cpu->node mapping.
With a typical setting of CPU_SETSIZE=1024, that's 8K.
If NUMA is not enabled, this array is not used, but it still consumes memory
(all 8K if data is mapped as huge pages and 4K otherwise (since the array is not aligned)).
Don't allocate the array at all, if NUMA is not enabled.

PiperOrigin-RevId: 565656252
Change-Id: Ie0dd2e5d87166e2da71bba8119a51bf75dcf32ab