tcmalloc: make TcmallocSlab::Push/PopBatch faster
Use cached slab pointer in Push/PopBatch as well.

We don't need to retry them if the slab pointer is not cached,
that's the behavior CpuCache wants -- if the thread was rescheduled,
CpuCache overflow/underflow needs to stop.
As the result we can remove "are we still on the same CPU" checks
from CpuCache code.

Another nice side-effect is that this frees 1 register on x86,
so we don't need to save/restore r12 and associated SAVE/RESTORE_REG macros.

name                           old cpu/op   new cpu/op   delta
BM_new_delete_slow_path/8      60.7µs ± 1%  56.7µs ± 0%  -6.52%  (p=0.008 n=5+5)

PiperOrigin-RevId: 566477344
Change-Id: I239b5cf0b81d2a60ce3739b7ae12b13c4945a2cc