trivial: use default member initialization for PageStat
otherwise we are at the mercy of the caller to initialize it,
and the only call point does it by hand right now.

minor associated cleanups: the tests can use PageStats{} when
we are expecting all zeros and explicitly use FieldsAre when
asserting all of the elements. this also means they won't break
if somebody merges a bad operator==.

PiperOrigin-RevId: 571970426
Change-Id: Ia2dd1f19b1763a46053e56580ed58f3380972e1a