Rolling forward size class resize with ShrinkOtherCache fix.
We might try to shrink an empty cache (with begin = current = end). While we check in ResizeSizeClasses if the capacity is 0, there is a chance that capacity drains to zero when we do this check and when we actually try to shrink the cache.

This cl also add an assertion to check if count > 0 while releasing to backing cache.

PiperOrigin-RevId: 572596470
Change-Id: Ia750f2614ddf9226f54120c909f1df28328e1c20