Correct the test SpanUtilizationHistogram
The test, it seems, assumes that the each call RemoveRange results in objects
being allocated from a single span.  This is not necessary.  Objects can be
allocated from two spans.  We change the test to record span pointers and use
them for deciding which span is being allocated/deallocated from.

PiperOrigin-RevId: 572992571
Change-Id: I6fb8497e9e1d1346171a97f0f91237a6909de98a