Avoid reading many-object span sampled state in assertions.
Spans with multiple objects are mutated by the CentralFreeList as objects are
allocate/deallocated from the span.  The Span contains a bitfield that shares
its byte with nonempty_index_ (mutated by CentralFreeList) and sampled_,
leading to a data race.

This assertion is superfluous, as we assert that the PageMap has a non-zero
size class for the object consistent with the one provided.  We also confirm
that sampled span do not have non-zero size classes upon
allocation/deallocation.

We add a stress test as well, for GetAllocatedSize, since this exercises a path
that checks sampled(), but only on large allocations.

PiperOrigin-RevId: 573323553
Change-Id: I0ab4ae1ff63130318113d397495f685aa0ec12dc