Use absolute number of allocations for span index in central freelist when objects_per_span is sufficiently low
Currently spans in the central free list are ordered based on the log of the number of allocations made from the spans.  Thus spans with allocations between two given powers of 2 are mapped to the same index.  When the number of objects per span is  <= 1 + kNumLists, we can instead use the absolute number of allocations for indexing.  This will spread the spans out and may result in lower fragmentation.

For unit testing, we added another param to MockStaticForwarder, the FakeCentralFreelistEnvironment and the CentralFreelistTest class.  The new param decides whether the new prioritization is in use.  The earlier param, class size, has been provided with more values (1024, 704) for testing more configs.  Unit test for testing span indexing has been modified for testing spans of various objects sizes.

PiperOrigin-RevId: 574917970
Change-Id: I05f156c4f5827d5f06097f55f7b7c56c932a5de7