Prepare for configurable max per-cpu capacity.
Currently we have very simple hard-coded max per-cpu per-class capacity:
 - 2048 objects for size class <= 10
 - 152 objects for larger size classes

While it's reasonable on high-level, it's too simplistic and too coarse-grained.
Some of the smaller classes are not as active as others,
and there are very active larger classes as well.

Prepare for externally configurable max capacities.
For now the old behavior is preserved.

PiperOrigin-RevId: 581170836
Change-Id: I1a5de2946b52a26f018177fc48286bc8eb5329c3