Refactor Span bitmap/intrusiveness methods.
Fix a number of small things:
1. Unexport a bunch of methods in Span.
2. Unify intrusiveness check. We have both IsIntrusive and open-coded
version of it in cold classes computation.
3. Reduce intrusiveness check to using bitmap. Currently we also check
for count for kCacheSize. This look excessive. We can't have fewer than
kCacheSize objects yet not using bitmap. That would mean we have unhealty
large cache (which makes bitmap excessive), or extremely small bitmap
(which is useless).
4. Remove GetBitmapSize. It just forwards to GetBitmapSizeHelper w/o adding
any logic, so no point in having both. But in the end GetBitmapSizeHelper
with all of the wrapping just duplicates the bitmap size computation logic,
so we don't get any encapsulation benefits. None of these are needed.
5. Rename !IsLessThanBitmapMinObjectSize to UseBitmapForSize.
The former feels too long, too low level and hard to parse.

PiperOrigin-RevId: 592580961
Change-Id: I816da9bc934649a3a11bad1f529676bb7a7d1747