Move Span methods.
1. Move FreelistPush definition from the class definition.
We don't have any other methods inline in the Span
(inlining methods makes it harder to read Span definition).

2. Move ListPopBatch definition from header file to source file.
I think we only want to have Push in the header, since it's relatively
small and has some chances of being inlined in CFL.
We don't want to inline ListPopBatch.
Moreover FreelistPopBatch that calls it is in the source file already,
so ListPopBatch in the header is pointless.

PiperOrigin-RevId: 592884105
Change-Id: Ib6013920d9592f43ee74eb22fda0071bf6200f15