Grow current cpu capacity only if wasn't rescheduled.
The intention is to grow the current cpu capacity after overflow/underflow,
if we are still running on the same cpu and wasn't rescheduled.
However, currently we can grow even if was rescheduled (but still on the same cpu,
which is likely).

This change fixes this by adding a primitive to store a value only
if wasn't rescheduled (still have the slab pointer cached).

As a side effect this also reduces code size, moves rseq primitive
from .S file to .cc file, which is I think the right thing long term.
Also prepares to subsequent changes, where this will be required
to simplify synchronization protocol for remote operations.

PiperOrigin-RevId: 601853833
Change-Id: I7c7433aeb857deb456c32e22e0b1b1afa694de39