Shorten jumps on new/delete fast paths.
Currently compiler emits a long jump to the sampling slow-path in new:

00000000005164c0 <TCMallocInternalNew>:
  ...
  51631b: 64 48 29 04 25 d0 fc ff ff    subq    %rax, %fs:-0x330
  516324: 0f 82 b6 44 00 00             jb      0x51a7e0 <sampling slow path>
  ...

It may increase load on CPU front end and I observed this to be slower
than short jump in memcpy code.

Use asm("") before slow path calls to prevent compiler from doing it.
With this change the code looks like:

00000000005164c0 <TCMallocInternalNew>:
  ...
  5164db: 64 48 29 04 25 d0 fc ff ff    subq    %rax, %fs:-0x330
  5164e4: 72 6c                         jb      0x516552 <_Znwm+0x92>
  ...
  ...
  ...
  516552: e9 e9 44 00 00                jmp     0x51aa40 <sampling slow path>

Sampling and other slow paths are orders of magnitude slower and less important
than the fast path. So it's reasonable to move any potential penalties
to the slow path cases.

PiperOrigin-RevId: 603295853
Change-Id: I2f9172e88c74bbf892635a49aabfd76b7011b2b8