Dedup cpu slab shrinking.
Shinking of slab capacity is duplicated between intra-cpu and inter-cpu resizing.
This produces lots of unnecessary cognitive load when reading and understanding
this code. It's hard/impossible to understand if the copies are exact, or there
are some small subtle differences somewhere. And this needs to be done by every
reader of this code each time.
Dedup the code.
If/when we tune these operations to be different, it's probably still better
to have a single version with a flag that will control the difference,
because the bulk of the logic will still be the same.

PiperOrigin-RevId: 603337557
Change-Id: Iac97ba35b5fff369e89df446dc18456d001554c0