Deflake startup_size_test.
Currently the test fails in 10+% of runs.
The largest value I observed in runs is:
Expected: (meta) <= (metadata_limit), actual: 31981494 vs 25165824
So bump the expectation from 20MB to 28MB.
This is quite an increase, but only this value made --runs_per_test=10000 pass.
If we want a lower value, we probably need to complicate this test a lot:
start 1000 subprocesses, aggregate values from them, and check median/max.
PiperOrigin-RevId: 604304554
Change-Id: I564d8e2fe8c5d048627f15df07406c76b0ec9d21