Add TCMALLOC_RELEASE_INLINE macro.
TCMALLOC_RELEASE_INLINE marks functions that need to be inlined in release
builds for performance reasons.

Gcc wants always_inline to be combined with inline, otherwise complains:
  error: 'always_inline' function might not be inlinable
which is easy miss the first time.

In debug builds we don't use always_inline to not excessively bloat
runtime and test functions, and to make breakpoints work. But we still use
inline to allow seamless use on functions in headers.

The macro is also 12 characters shorter than "inline ABSL_ATTRIBUTE_ALWAYS_INLINE",
which helps to minimize line wraps while still keeping the code
80-column punchcard friendly.

PiperOrigin-RevId: 604581455
Change-Id: I0a33a810543e26d8813991678db6419603635c36