Improve and simplify CpuCache reclaim logic.
Currently we look at the number of misses and bytes used by a CPU
to detect when it's unused. This is both complex and imprecise.
It's unlikely that misses/used bytes stay the same when a CPU cache is used,
but not impossible to imagine as well. If a thread allocates and frees
N objects using existing capacity, then both misses/used bytes stay the same.

Thanks for slab pointer caching, we can now record the exact fact
of a CPU cache being used without relying on the indirect signals.

PiperOrigin-RevId: 604595674
Change-Id: I6aed32a8697ebd0b70e2f7f05afe025f658836e7