Revert: Improve and simplify CpuCache reclaim logic.
Currently we look at the number of misses and bytes used by a CPU
to detect when it's unused. This is both complex and imprecise.
It's unlikely that misses/used bytes stay the same when a CPU cache is used,
but not impossible to imagine as well. If a thread allocates and frees
N objects using existing capacity, then both misses/used bytes stay the same.

PiperOrigin-RevId: 604657753
Change-Id: I052c9da091fe709e5973101e081e59e1c9fbba37