Revert: Fix CpuCache destruction leaks.
TcmallocSlab::Destroy deletes the current slabs,
but since we can have more slabs with slab resizing,
when CpuCache is destroyed we leak all of the remaining ones.

Instead of deleting one in TcmallocSlab::Destroy and returning the pointer,
don't delete any in TcmallocSlab::Destroy and let caller delete all of them.

PiperOrigin-RevId: 604658413
Change-Id: Ib14404d61df4a476cd10918fa85b9aa8e8dc122a