Revert: Add TCMALLOC_RELEASE_INLINE macro.
TCMALLOC_RELEASE_INLINE marks functions that need to be inlined in release
builds for performance reasons.

Gcc wants always_inline to be combined with inline, otherwise complains:
  error: 'always_inline' function might not be inlinable
which is easy miss the first time.

In debug builds we don't use always_inline to not excessively bloat
runtime and test functions, and to make breakpoints work. But we

PiperOrigin-RevId: 604668021
Change-Id: I0cfd90f59bd72157dcba0c230d24b797188449ff