4-byte slab headers.
Currently the slab header is 8 bytes and holds [begin, current, end] indices.
This change switches to the new scheme where the header is 4 bytes and
holds only [current, end] indices. Begin indices are stored in a separate array
and are used only in slower operations.

In the new scheme we mark the stub element stored before the size class range
(the one that we use for prefetching when we pop the last element) with
low bit set.

PiperOrigin-RevId: 609982618
Change-Id: I7cff2c196c6a52f5fb7b08f1548c3e72a5ae2ad9