Improve and simplify guarded sampling.
Currently we have sampling hysteresis controlled by striving_to_guard_.
With the default sampling rate (1/50), the hysteresis is 20%:
we stop sampling when we drop below 1/50 and start sampling again only at 1/60.

There is only explanation as to "what" but not "why",
I did not find any in the original commit either.

The problem with 20% hysteresis is that we stop sampling for 20% of the current
program uptime. E.g. if uptime is 5 days, we stop sampling for 1 day (like at all).
I don't see how this is the behavior we want. Remove the hysteresis.

Without the hysteresis we start looking for suitable guarded allocations
after ~50 sampled allocations (look at every). Once we find one that passes
the stack trace filter, we guard it.
Given that sampled allocations are already randomized (and the application
behavior is random enough on its own) + stack trace filter that is supposed to saturate,
this looks like a reasonable behavior and better than stopping sampling for days.

This may also contribute to some of the flakiness of the memory_errors_test.

PiperOrigin-RevId: 611344713
Change-Id: If5ff7c0afcde8a9f4a4dcdd7f374d59a229e16c6