Move size reciprocal from Span to CFL.
We store properties that are common to a size class (object size, objects per span)
in CFL rather than in the Span. Move size reciprocal to CFL as well.
This can help to reduce Span size to 32 bytes in future and/or
increase bitmap size to 128 bits.
This also resolves the old TODO about computing reciprocal once per size class
rather than for each Span object.

PiperOrigin-RevId: 614948333
Change-Id: Icc53060799d0e436d546aafe17eeea4d60e9c759