Use more virtual address space for heap.
Currently we reduce available virtual address space for heap because of:

  //  *  One additional bit below that, to avoid collisions with mappings that
  //     tend to be placed in the upper half of the address space (e.g. stack,
  //     executable, and VDSO mappings).

Since we don't use MAP_FIXED, this is not a valid reason to do this.
There is nothing wrong with trying to mmap at already occupied address,
we try to do it anyway (we don't know where all existing mappings are).

testing/system-alloc_test.cc tries to allocate 1<<kTagShift.
This starts breaking with the proposed SelSan mapping (and may be breaking
today with NUMA support, not sure). The proposed SelSan mapping makes tags
stable with/without NUMA, so there is 1 less bit for tag ranges w/o NUMA as well.
For sampled allocations of size 1<<kTagShift (0x40000000000),
there are only 2 suitable addresses on x86_64: 0 and 0x200000000000.
0 is not a legit address for mmaps and rejected by sane kernels.
0x200000000000 may be already taken if there was at least 1 sampled
allocation before the test has started.
So we end up with no suitable ranges for a sampled allocation of size 1<<kTagShift.

With the additional bit we have 2 more suitable ranges for 1<<kTagShift allocation.

PiperOrigin-RevId: 615365560
Change-Id: I0d744e5f396cf6ef19e00ee8fc704e4e2bbab4dc