Configure experiment for larger spans.
This combines two optimizations:
* Spans can be formally aligned to cachelines.
* With the extra space, we can have a larger Span-local cache when using the
  compressed linked list representation.

The allocation fast path impact in the CentralFreeList is limited to reading a
global variable on push to know the available maximum span size.  This change applies to both experiment and control arms and would be removed upon conclusion of experiments.

Pop reads the cache size from the span that was previously stored by BuildFreelist/Push.

PiperOrigin-RevId: 617293681
Change-Id: I8e65ca139c99f3b9768bedba5c3c0a182e995061