Fix MallocExtension exposure of experiments under sanitizers.
Currently all MallocExtension properties, including experiments,
are provided by the weak function in tcmalloc.cc.
This file is not compiled under sanitizers, but experiments are still
enabled and exposed under sanitizers as well.
As the result from the experiment.h interface it looks like experiments
are enabled, but from MallocExtension interface it looks like they are not.
This may lead to arbitrary misbehavior, some code may conclude an experiment
is enabled, but another code will conclude it's not.
Fix this by moving things to where they belong.

PiperOrigin-RevId: 617540658
Change-Id: I0620bd7883985f47becff3ffe4be1579e38860a6