Fix the logic for randomly enabling experiments.
We found that it is possible that hashes can be such that no experiment is enabled.  This results in a failure since the test checks that at least one experiment is enabled.  So in case the current logic fails to select any experiment, we enable  half the experiments by alternating between the experiments enabled and disabled.

PiperOrigin-RevId: 623210716
Change-Id: Icc17444c46c0087219434acc37e335eb0433218f