Speed up selsan tag setting.
Currently compiler replaces SetTag with a memset call.
Since the call is a non-tail call, this introduces a stack frame
into operator delete.
Use a custom code that is effectively an inlined memset.
This removes the stack frame and slightly optimizes size rounding up
(we don't need to do it separately, can combine with other calculations).

PiperOrigin-RevId: 623389451
Change-Id: Ife220752ad19c436e8ebf8b3a7747f62fb534cfc