Sort out naming real vs. virtual CPUs
Various helpers have similar names, referring to either real CPU IDs or virtual
CPU IDs. Attempt to sort out naming, so that we refer either to a "real CPU ID"
or a "virtual CPU ID" (which is an index into per-CPU slabs).

Where it made sense, the naming was adjusted to be more intuitive and less
redundant:

 * "flat vCPUs" -> "rseq vCPUs", since "vCPU" itself implies "flatness"
 * GetCurrent*() -> "current" is implied, and is dropped
 * VirtualRseqCpuId -> RseqVirtualCpuId (subsystem first)
 * VirtualUserCpuId -> UserVirtualCpuId (counterpart to RseqVirtualCpuId)

NFC.

PiperOrigin-RevId: 627395415
Change-Id: I54c95b7bf6d34744fa42006c2d7663cccec45ea6