Constrain vCPU ID interface and introduce vCPU caching
Allow caching of the vCPU value in thread-local storage. This is done by constraining the interface with which TCMalloc code may access vCPUs (with static class VirtualCpu).

Indeed, without revalidating the slab pointer (or any other state) derived from the vCPU ID in a RSEQ critical section, simply using the vCPU ID used to derive the slab pointer is safe, because a thread may be preempted at any point and reading the vCPU ID can be any value.

This change also exposed subtle inconsistencies in the DeallocationProfiler, which simply used the latest vCPU ID, which may have already changed due to preemption. By using the cached vCPU ID, the results from the DeallocationProfiler should be more accurate.

For better consistency, the API in percpu.h is constrained as much as possible.

PiperOrigin-RevId: 627637423
Change-Id: I0b5be006fbcffc2381ef429032c8c47bffbb5897