Simulate reentrancy under the page heap lock with fuzzing.
The page heap may release the page heap lock when we are returning memory to
the OS.  While we do not differentiate this from (deprecated/being cleaned up)
lockful releases, we can examine whether pageheap_lock.IsHeld() is true.  If it
is not, we can use a previous snippet of our fuzzer dsl to drive recursively
modifying the page heap.  This can let us simulate multi-threaded interleavings
of release with further allocations.

PiperOrigin-RevId: 631461361
Change-Id: I41a2cb165d369cf20341e689e8326a8106918a5a