Pass the hot_cold_t hint value to Sample{Large,Small}Allocation.
Instead of using the value from CppPolicy, store the actual value of the hint passed to operator new. This allows us to distinguish between different values passed rather than the decision of whether to access as hot or cold. The actual decision can be inferred from the "access_allocated" tag in the profile.

PiperOrigin-RevId: 639101074
Change-Id: Ifde25ec782cb29df4e1e98863d9611ecbccde59f