GPA: Add TrySample benchmark
Add a benchmark that exclusively benchmarks the guarded sampling-decision
overhead (excluding the overhead of making a guarded allocation). Since the
decision to guard may change, the only way to fairly compare differences in the
sampling decision algorithm is to exclude the overhead of allocations.

The benchmark simply exhausts the pool before the profiled portion of the code.

PiperOrigin-RevId: 641949319
Change-Id: I2a7159cf367b9565c9847c8a36a3fe24de9cfd17